Update activityDetailPage style
