+++ packages/inside/src/PageContainer/Club/index.ts
@@ -1,3 +1,2 @@
 export { default as ActivityCreatePage } from './ActivityCreatePage'
-export { default as ActivityDetailPage } from './ActivityDetailPage'
 export { default as ActivityModifyPage } from './ActivityModifyPage'
