+++ packages/inside/src/app/main/club/[clubId]/student/[studentId]/activity/page.tsx
@@ -1,5 +1,5 @@
 import { StudentIdProps } from '@bitgouel/types'
-import { ActivityListPage } from '@bitgouel/common/src/pages'
+import { ActivityListPage } from '@bitgouel/common'
 
 const ActivityList = ({ params }: { params: StudentIdProps }) => {
   return <ActivityListPage studentIdProps={params} />
