+++ packages/outside/src/app/main/club/[clubId]/student/[studentId]/activity/[activityId]/page.tsx
@@ -1,15 +1,11 @@
 import { ActivityDetailPage } from '@bitgouel/common'
 import { StudentIdProps } from '@bitgouel/types'
 
-interface Props {
-  params: {
-    activityId: string
-    studentId: string
-    clubId: string
-  }
-}
-
-const ActivityDetail: React.FC<Props> = ({ params }) => {
+const ActivityDetail = ({
+  params,
+}: {
+  params: StudentIdProps & { activityId: string }
+}) => {
   return <ActivityDetailPage studentIdProps={params} />
 }
