+++ shared/common/src/pages/activity/detail/style.ts
