--- shared/common/src/pages/club/index.ts
@@ -1,2 +1,0 @@
-export {default as ClubDetailPage} from './detail'
-export {default as StudentPage} from './student'
