+++ shared/common/src/pages/club/student/index.tsx
@@ -1,4 +1,5 @@
 'use client'
+
 import {
   TokenManager,
   useGetCertificateList,
