+++ shared/common/src/pages/index.ts
@@ -1,7 +1,6 @@
 export { default as LoginPage } from './login'
 export { default as SignUpPage } from './signUp'
 export { default as HomePage } from './home'
-export { ClubDetailPage, StudentPage } from './club'
 export {
   PostPage,
   PostDetailPage,
@@ -14,4 +13,7 @@ export {
 export { default as MyPage } from './my'
 export { InquiryPage, InquiryDetailPage } from './inquiry'
 export { LecturePage, LectureDetailPage } from './lecture'
-export { ActivityListPage } from './activity'
+export { default as ActivityListPage } from './activity/list'
+export { default as ActivityDetailPage } from './activity/detail'
+export { default as ClubDetailPage } from './club/detail'
+export { default as StudentPage } from './club/student'
